move Neon impl as static inline in header to enable inlining
